+++ awsx/lb/applicationLoadBalancer.ts
@@ -40,6 +40,7 @@ export class ApplicationLoadBalancer extends schema.ApplicationLoadBalancer {
       subnetIds,
       subnets,
       defaultTargetGroup,
+      defaultTargetGroupPort,
       defaultSecurityGroup,
       listener,
       listeners,
@@ -166,6 +167,12 @@ export class ApplicationLoadBalancer extends schema.ApplicationLoadBalancer {
       parent: this,
     });
 
+    if (defaultTargetGroup !== undefined && defaultTargetGroupPort !== undefined) {
+      throw new Error(
+        "Only one of `defaultTargetGroup` or `defaultTargetGroupPort` can be provided.",
+      );
+    }
+
     const defaultProtocol = getDefaultProtocol(args);
 
     this.defaultTargetGroup = new aws.lb.TargetGroup(
@@ -242,11 +249,12 @@ function getDefaultProtocol(
       return undefined;
     }
   }
-  return getListenerProtocol(listener);
+  return getListenerProtocol(listener, args.defaultTargetGroupPort);
 }
 
 function getListenerProtocol(
   listener: schema.ListenerInputs | undefined,
+  defaultPort?: pulumi.Input<number>,
 ): { port: pulumi.Input<number>; protocol: pulumi.Input<string> } | undefined {
   if (listener) {
     const { port, protocol } = listener;
@@ -266,7 +274,7 @@ function getListenerProtocol(
       };
     }
   }
-  return { port: 80, protocol: "HTTP" };
+  return { port: defaultPort ?? 80, protocol: "HTTP" };
 }
 
 function portToProtocol(port: number) {